Rename to CreatingResourcesIntegrationTest
